CIB-2404: Switch mark for clean build to use persistent agent messaging.  Removed the force clean agents table, and changed the clean action to send out a persistent message to delete the directories.  For this I needed to extend the delete task to allow patterns (as persistent work directories are configurable), which in turn required a type squeezer for maps.

+++ com.zutubi.pulse.servercore/src/java/com/zutubi/pulse/servercore/agent/DeleteDirectoryTask.java
@@ -1,8 +1,17 @@
 package com.zutubi.pulse.servercore.agent;
 
+import com.zutubi.pulse.servercore.ServerRecipePaths;
+import com.zutubi.pulse.servercore.bootstrap.ConfigurationManager;
 import com.zutubi.pulse.servercore.cleanup.FileDeletionService;
+import com.zutubi.tove.variables.GenericVariable;
+import com.zutubi.tove.variables.HashVariableMap;
+import com.zutubi.tove.variables.VariableResolver;
+import com.zutubi.tove.variables.api.ResolutionException;
+import com.zutubi.tove.variables.api.VariableMap;
+import com.zutubi.util.logging.Logger;
 
 import java.io.File;
+import java.util.Map;
 
 /**
  * A synchronisation task that deletes a directory.  If the directory does not
@@ -13,7 +22,12 @@ import java.io.File;
  */
 public class DeleteDirectoryTask implements SynchronisationTask
 {
-    private String path;
+    private static final Logger LOG = Logger.getLogger(DeleteDirectoryTask.class);
+    
+    private String agentDataDirectoryPattern;
+    private String pathPattern;
+    private Map<String, String> variables;
+    private transient ConfigurationManager configurationManager;
     private transient FileDeletionService fileDeletionService;
 
     public DeleteDirectoryTask()
@@ -23,20 +37,61 @@ public class DeleteDirectoryTask implements SynchronisationTask
     /**
      * Creates a task to delete a directory.
      *
-     * @param path path of the directory to delete
+     * @param agentDataDirectoryPattern pattern specifying the location of the
+     *                                  agent data directory
+     * @param pathPattern pattern specifying the directory to delete; may
+     *                    include references to variables passed to this
+     *                    constructor and/or $(data.dir), which is provided on
+     *                    the agent before resolving the variables
+     * @param variables   variables used to resolve the path patterns (excludes
+     *                    the data directory, which is known on the agent side)
      */
-    public DeleteDirectoryTask(String path)
+    public DeleteDirectoryTask(String agentDataDirectoryPattern, String pathPattern, Map<String, String> variables)
     {
-        this.path = path;
+        this.agentDataDirectoryPattern = agentDataDirectoryPattern;
+        this.pathPattern = pathPattern;
+        this.variables = variables;
     }
 
     public void execute()
     {
-        File dir = new File(path);
-        if (dir.exists())
+        String path = resolvePattern();
+        if (path != null)
         {
-            fileDeletionService.delete(dir);
+            File dir = new File(path);
+            if (dir.exists())
+            {
+                fileDeletionService.delete(dir);
+            }
         }
+    }
+
+    private String resolvePattern()
+    {
+        VariableMap map = new HashVariableMap();
+        for (Map.Entry<String, String> entry: variables.entrySet())
+        {
+            map.add(new GenericVariable<String>(entry.getKey(), entry.getValue()));
+        }
+        
+        map.add(new GenericVariable<String>(ServerRecipePaths.PROPERTY_DATA_DIR, configurationManager.getUserPaths().getData().getAbsolutePath()));
+        try
+        {
+            map.add(new GenericVariable<String>(ServerRecipePaths.PROPERTY_AGENT_DATA_DIR, VariableResolver.resolveVariables(agentDataDirectoryPattern, map, VariableResolver.ResolutionStrategy.RESOLVE_NON_STRICT)));
+            return VariableResolver.resolveVariables(pathPattern, map);
+        }
+        catch (ResolutionException e)
+        {
+            // This should not happen with non-strict resolution.
+            LOG.debug(e);
+        }
+        
+        return null;
+    }
+
+    public void setConfigurationManager(ConfigurationManager configurationManager)
+    {
+        this.configurationManager = configurationManager;
     }
 
     public void setFileDeletionService(FileDeletionService fileDeletionService)
